+++ android/src/main/java/dev/barooni/capacitor/calendar/implementation/CapacitorCalendar.kt
@@ -36,6 +36,7 @@ import dev.barooni.capacitor.calendar.models.results.DeleteEventsByIdResult
 import dev.barooni.capacitor.calendar.models.results.GetDefaultCalendarResult
 import dev.barooni.capacitor.calendar.models.results.ListCalendarsResult
 import dev.barooni.capacitor.calendar.models.results.ListEventsInRangeResult
+import dev.barooni.capacitor.calendar.models.templates.PluginCompletion
 import dev.barooni.capacitor.calendar.utils.ImplementationHelper
 
 class CapacitorCalendar(
@@ -144,11 +145,15 @@ class CapacitorCalendar(
         return ListCalendarsResult(calendars)
     }
 
-    fun getDefaultCalendar(): GetDefaultCalendarResult {
-        val cr = plugin.context.contentResolver
-        val calendars = ImplementationHelper.listCalendars(cr)
-        val primaryCalendar = calendars.find { it.isPrimary == true }
-        return GetDefaultCalendarResult(primaryCalendar)
+    fun getDefaultCalendar(completion: PluginCompletion<GetDefaultCalendarResult>) {
+        try {
+            val cr = plugin.context.contentResolver
+            val calendars = ImplementationHelper.listCalendars(cr)
+            val defaultCalendar = ImplementationHelper.resolveDefaultCalendar(calendars)
+            completion(GetDefaultCalendarResult(defaultCalendar), null)
+        } catch (error: Exception) {
+            completion(null, error)
+        }
     }
 
     fun openCalendar(input: OpenCalendarInput) {